fix arrays and reimplement jump instructions

+++ TypeSizes.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include <cstdint>
 #include <cstddef>
-using ProgramAddressType = uint64_t;
+using ProgramAddressType = size_t;
 using NumberType = double;
 using IntegerType = int64_t;
 using UIntegerType = uint64_t;
